+++ include/_node_parser.h
@@ -45,9 +45,9 @@ namespace cfg::parser
      */
     [[nodiscard]] static auto parse(const cfg::reader::SectionReader& node_reader,
                                     std::istream& mesh_stream,
-                                    const Mode mode)
+                                    const reader::Mode mode)
     {
-      if (mode == Mode::BINARY)
+      if (mode == reader::Mode::BINARY)
       {
         mesh_stream.ignore(1);  // Skip spare char
       }
@@ -55,10 +55,10 @@ namespace cfg::parser
       {
         NodeHeader node_header{};
 
-        node_header.n_blocks = read_one<size_t>(node_reader, mesh_stream, mode);
-        node_header.n_nodes  = read_one<size_t>(node_reader, mesh_stream, mode);
-        node_header.min_tag  = read_one<size_t>(node_reader, mesh_stream, mode);
-        node_header.max_tag  = read_one<size_t>(node_reader, mesh_stream, mode);
+        node_header.n_blocks = reader::read_one<size_t>(node_reader, mesh_stream, mode);
+        node_header.n_nodes  = reader::read_one<size_t>(node_reader, mesh_stream, mode);
+        node_header.min_tag  = reader::read_one<size_t>(node_reader, mesh_stream, mode);
+        node_header.max_tag  = reader::read_one<size_t>(node_reader, mesh_stream, mode);
 
         return node_header;
       }();
@@ -91,9 +91,9 @@ namespace cfg::parser
      * @param environment Contains the calling environment, in particular passes the Parallel field
      * @returns The node vector.
      */
-    [[nodiscard]] static auto parse(const cfg::reader::SectionReader& node_reader,
+    [[nodiscard]] static auto parse(const reader::SectionReader& node_reader,
                                     std::istream& mesh_stream,
-                                    const Mode mode,
+                                    const reader::Mode mode,
                                     const NodeHeader& node_header,
                                     const NodeEnvironment& environment)
     {
@@ -140,14 +140,14 @@ namespace cfg::parser
      *          parametric and the number of nodes in the block.
      */
     [[nodiscard]] static std::tuple<int, int, bool, size_t> parse_node_block_header(
-        const cfg::reader::SectionReader& node_reader,
+        const reader::SectionReader& node_reader,
         std::istream& mesh_stream,
-        const Mode mode) noexcept
+        const reader::Mode mode) noexcept
     {
-      const auto block_dim   = read_one<int>(node_reader, mesh_stream, mode);
-      const auto block_tag   = read_one<int>(node_reader, mesh_stream, mode);
-      const auto block_param = read_one<int>(node_reader, mesh_stream, mode);
-      const auto block_nodes = read_one<size_t>(node_reader, mesh_stream, mode);
+      const auto block_dim   = reader::read_one<int>(node_reader, mesh_stream, mode);
+      const auto block_tag   = reader::read_one<int>(node_reader, mesh_stream, mode);
+      const auto block_param = reader::read_one<int>(node_reader, mesh_stream, mode);
+      const auto block_nodes = reader::read_one<size_t>(node_reader, mesh_stream, mode);
 
       return {block_dim, block_tag, bool{static_cast<bool>(block_param)}, block_nodes};
     }
@@ -164,13 +164,13 @@ namespace cfg::parser
     [[nodiscard]] static std::vector<size_t> parse_node_idx(const cfg::reader::SectionReader& node_reader,
                                                             const size_t block_nodes,
                                                             std::istream& mesh_stream,
-                                                            const Mode mode) noexcept
+                                                            const reader::Mode mode) noexcept
     {
       std::vector<size_t> indices(block_nodes);
 
       for (size_t node = 0; node < block_nodes; node++)
       {
-        indices[node] = read_one<size_t>(node_reader, mesh_stream, mode);
+        indices[node] = reader::read_one<size_t>(node_reader, mesh_stream, mode);
       }
 
       return indices;
@@ -189,13 +189,13 @@ namespace cfg::parser
         const cfg::reader::SectionReader& node_reader,
         const size_t block_nodes,
         std::istream& mesh_stream,
-        const Mode mode) noexcept
+        const reader::Mode mode) noexcept
     {
       std::vector<std::array<double, 3>> coords(block_nodes);
 
       auto pop_component = [&node_reader, &mesh_stream, mode]() -> double
       {
-        return read_one<double>(node_reader, mesh_stream, mode);
+        return reader::read_one<double>(node_reader, mesh_stream, mode);
       };
 
       for (size_t node = 0; node < block_nodes; node++)
@@ -247,7 +247,7 @@ namespace cfg::parser
    * @param parallel The parallel environment.
    * @returns A function to read nodes from a GMSH file.
    */
-  std::function<std::vector<Node<3>>(const cfg::reader::SectionReader&, std::ifstream&, const Mode)> make_node_reader(
+  std::function<std::vector<Node<3>>(const cfg::reader::SectionReader&, std::ifstream&, const reader::Mode)> make_node_reader(
       const cfg::utils::Parallel& parallel);
 }  // namespace cfg::parser