+++ include/node_parser.h
@@ -8,41 +8,13 @@
 
 #include <array>
 #include <cstddef>
-#include <istream>
 
+#include <reader.h>
 #include <section_reader.h>
 #include <utils.h>
 
 namespace cfg::parser
 {
-  /**
-   * Identifies which mode the mesh file is read in: ASCII (formatted) or BINARY (unformatted).
-   */
-  enum class Mode
-  {
-    ASCII,
-    BINARY
-  };
-
-  /**
-   * Reads a single item from the reader, according to the mode.
-   */
-  template <class C>
-  [[nodiscard]] C read_one(const cfg::reader::SectionReader& node_reader, std::istream& mesh_stream, const Mode mode)
-  {
-    C val;
-    if (mode == Mode::ASCII)
-    {
-      node_reader(mesh_stream) >> val;
-    }
-    else
-    {
-      const size_t nchar = sizeof(C) / sizeof(char);
-      mesh_stream.read((char*)&val, nchar);
-    }
-    return val;
-  }
-
   /**
    * A mesh node of arbitrary dimension `d`. This stores the node's index and coordinates.
    */
@@ -66,10 +38,10 @@ namespace cfg::parser
   auto read_X(const H hdr_parser, const D data_parser, const E environment, const V validator)
   {
     return [hdr_parser, data_parser, environment, validator](
-               const cfg::reader::SectionReader& section_reader, std::ifstream& mesh_stream, const Mode mode)
+               const reader::SectionReader& section_reader, std::ifstream& mesh_stream, const reader::Mode mode)
     {
-      const auto hdr       = hdr_parser.parse(section_reader, mesh_stream, mode);
-      const auto data      = data_parser.parse(section_reader, mesh_stream, mode, hdr, environment);
+      const auto hdr  = hdr_parser.parse(section_reader, mesh_stream, mode);
+      const auto data = data_parser.parse(section_reader, mesh_stream, mode, hdr, environment);
       validator.validate(data, hdr);
       return data;
     };
@@ -82,7 +54,9 @@ namespace cfg::parser
    * @param mode        Flag indicating whether the file was opened in ASCII or binary mode.
    * @param parallel    The parallel environment.
    */
-  void read_nodes(std::ifstream& mesh_stream, const Mode mode, const cfg::utils::Parallel& parallel);
+  [[nodiscard]] std::vector<Node<3>> read_nodes(std::ifstream& mesh_stream,
+                                                const reader::Mode mode,
+                                                const cfg::utils::Parallel& parallel);
 }  // namespace cfg::parser
 
 #endif  // __CFG_NODE_PARSER_H_