+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 ### Changed
 
+- `read_nodes` returns the Nodes
 - Refactored the `read_nodes` function in terms of a generic `read_X` function
 - Replaced `parse_node_header` with `HeaderParser::parse`
 - Replaced `parse_node_blocks` with `DataParser::parse`